@@ -1,10 +1,11 @@
 {
   "description": "Browsertime",
-  "version": "2.0.0-alpha1",
+  "version": "2.1.3",
   "bin": "./bin/browsertime.js",
   "dependencies": {
-    "adbkit": "2.8.1",
-    "alto-saxophone": "2.33.0",
+    "@sitespeed.io/throttle": "0.1.0",
+    "adbkit": "2.11.0",
+    "alto-saxophone": "2.35.0",
     "bluebird": "3.5.0",
     "chrome-har": "0.2.3",
     "execa": "0.6.1",
@@ -22,8 +23,7 @@
     "mkdirp": "0.5.1",
     "moment": "2.18.1",
     "selenium-webdriver": "3.6.0",
-    "@sitespeed.io/throttle": "0.1.0",
-    "soprano-saxophone": "0.18.0",
+    "soprano-saxophone": "0.19.1",
     "valid-url": "1.0.9",
     "xvfb": "0.2.3",
     "yargs": "9.0.1"
@@ -31,16 +31,16 @@
   "devDependencies": {
     "chai": "^4.1.2",
     "chai-as-promised": "~7.1.1",
-    "eslint": "~4.2.0",
-    "eslint-config-prettier": "~2.3.0",
-    "eslint-plugin-prettier": "~2.1.2",
+    "eslint": "^4.11.0",
+    "eslint-config-prettier": "^2.7.0",
+    "eslint-plugin-prettier": "^2.3.1",
     "longjohn": "0.2.11",
     "mocha": "~3.1.2",
-    "prettier": "~1.5.3",
+    "prettier": "^1.8.2",
     "sequencediagrams": "0.0.8"
   },
   "engines": {
-    "node": ">=8.7.0"
+    "node": ">=8.9.0"
   },
   "main": "index.js",
   "scripts": {
