@@ -1,11 +1,10 @@
 'use strict';
 
-const Promise = require('bluebird'),
-  log = require('intel'),
-  webdriver = require('selenium-webdriver'),
-  harBuilder = require('../../support/harBuilder'),
-  perflogParser = require('chrome-har'),
-  traceCategoriesParser = require('../../support/traceCategoriesParser');
+const log = require('intel');
+const { Type } = require('selenium-webdriver').logging;
+const harBuilder = require('../../support/harBuilder');
+const perflogParser = require('chrome-har');
+const traceCategoriesParser = require('../../support/traceCategoriesParser');
 
 const CHROME_NAME_AND_VERSION_JS = `return (function() {
   var match = navigator.userAgent.match(/(Chrom(e|ium))\\/(([0-9]+\\.?)*)/);
@@ -29,89 +28,62 @@ class ChromeDelegate {
     this.chrome = chrome;
   }
 
-  onStartRun() {
+  async onStartRun() {
     this.hars = [];
-    return Promise.resolve();
   }
 
-  onStartIteration(runner) {
+  async onStartIteration(runner) {
     if (this.skipHar) {
-      return Promise.resolve();
+      return;
     }
 
     // remove irrelevant entries from performance log
     // and since Chromedriver 2.29 there's a bug
     // https://bugs.chromium.org/p/chromedriver/issues/detail?id=1811
     // that can be fixed by emptying the logs twice :|
-    return runner
-      .getLogs(webdriver.logging.Type.PERFORMANCE)
-      .then(() => runner.getLogs(webdriver.logging.Type.PERFORMANCE));
+    await runner.getLogs(Type.PERFORMANCE);
+    await runner.getLogs(Type.PERFORMANCE);
   }
 
-  onStopIteration(runner, index, results) {
+  async onStopIteration(runner, index, results) {
     if (this.skipHar) {
-      return Promise.resolve();
+      return;
     }
 
     log.debug('Getting performance logs from Chrome');
 
-    return (
-      runner
-        .getLogs(webdriver.logging.Type.PERFORMANCE)
-        // need timeout to avoid hanging forever in case of error
-        .timeout(90000, 'Extracting Chrome performance log took too long.')
-        .map(entry => JSON.parse(entry.message).message)
-        .tap(messages => {
-          if (this.logPerfEntries) {
-            results.extraJson['chromePerflog-' + index + '.json'] = messages;
-          }
-        })
-        .tap(messages => {
-          if (this.collectTracingEvents) {
-            results.extraJson[
-              'trace-' + index + '.json'
-            ] = traceCategoriesParser.parse(messages);
-          }
-        })
-        .then(messages => perflogParser.harFromMessages(messages))
-        .then(har =>
-          runner
-            .runScript(CHROME_NAME_AND_VERSION_JS, 'CHROME_NAME_AND_VERSION_JS')
-            .then(browserInfo => {
-              if (this.chrome.mobileEmulation) {
-                browserInfo.name =
-                  'Chrome Emulated ' + this.chrome.mobileEmulation.deviceName;
-              }
-              return harBuilder.addBrowser(
-                har,
-                browserInfo.name,
-                browserInfo.version
-              );
-            })
-        )
-        .then(har => {
-          if (har.log.pages.length > 0) {
-            return runner
-              .runScript(PAGE_TITLE_JS, 'PAGE_TITLE_JS')
-              .then(title => {
-                har.log.pages[0].title = title + ' run ' + (index + 1);
-                return har;
-              });
-          } else {
-            // FIXME this is to avoid a crash, but should have better error handling in this case.
-            return har;
-          }
-        })
-        .then(har => {
-          this.hars.push(har);
-        })
+    const logs = await runner.getLogs(Type.PERFORMANCE);
+
+    const messages = logs.map(entry => JSON.parse(entry.message).message);
+
+    if (this.logPerfEntries) {
+      results.extraJson[`chromePerflog-${index}.json`] = messages;
+    }
+    if (this.collectTracingEvents) {
+      const trace = traceCategoriesParser.parse(messages);
+      results.extraJson[`trace-${index}.json`] = trace;
+    }
+
+    const har = perflogParser.harFromMessages(messages);
+    const info = await runner.runScript(
+      CHROME_NAME_AND_VERSION_JS,
+      'CHROME_NAME_AND_VERSION_JS'
     );
+    if (this.chrome.mobileEmulation) {
+      info.name = `Chrome Emulated ${this.chrome.mobileEmulation.deviceName}`;
+    }
+    harBuilder.addBrowser(har, info.name, info.version);
+
+    if (har.log.pages.length > 0) {
+      const title = await runner.runScript(PAGE_TITLE_JS, 'PAGE_TITLE_JS');
+      har.log.pages[0].title = `${title} run ${index + 1}`;
+    }
+
+    this.hars.push(har);
   }
 
-  onStopRun(result) {
+  async onStopRun(result) {
     result.har = harBuilder.mergeHars(this.hars);
-
-    return Promise.resolve();
   }
 }
 
