@@ -1,29 +1,24 @@
 'use strict';
 
-const Promise = require('bluebird'),
-  log = require('intel'),
-  BrowserError = require('../../support/errors').BrowserError,
-  remove = require('lodash.remove'),
-  harBuilder = require('../../support/harBuilder');
+const BrowserError = require('../../support/errors').BrowserError;
+const remove = require('lodash.remove');
+const harBuilder = require('../../support/harBuilder');
 
 class FirefoxDelegate {
   constructor({ skipHar = false }) {
     this.skipHar = skipHar;
   }
 
-  onStartRun() {
+  async onStartRun() {
     this.index = 1;
     this.hars = [];
-    return Promise.resolve();
   }
 
-  onStartIteration() {
-    return Promise.resolve();
-  }
+  async onStartIteration() {}
 
-  onStopIteration(runner) {
+  async onStopIteration(runner) {
     if (this.skipHar) {
-      return Promise.resolve();
+      return;
     }
 
     const script = `
@@ -47,44 +42,30 @@ class FirefoxDelegate {
             } else {
               triggerExport();
             }`;
-    return runner
-      .runAsyncScript(script, 'GET_HAR_SCRIPT')
-      .then(reply => {
-        if (reply.error) {
-          throw new BrowserError('Error in Firefox HAR generation', {
-            cause: reply.error
-          });
-        } else {
-          return reply.har;
-        }
-      })
-      .then(JSON.parse)
-      .then(har => {
-        har.log.pages[0].title += ' run ' + this.index;
 
-        // Firefox inlude entries in the HAR that are from the local cache
-        // and adds them with a time of 0.
-        // lets remove them.
-        remove(har.log.entries, function(entry) {
-          return entry.time === 0;
-        });
-        this.hars.push(har);
-        this.index++;
-      })
-      .catch(e => {
-        log.error(
-          "Couldn't generate the HAR from Firefox from the HAR Export Trigger."
-        );
-        throw e;
+    const harResult = await runner.runAsyncScript(script, 'GET_HAR_SCRIPT');
+    if (harResult.error) {
+      throw new BrowserError('Error in Firefox HAR generation', {
+        cause: harResult.error
       });
+    }
+    const har = JSON.parse(harResult.har);
+    har.log.pages[0].title += ' run ' + this.index;
+
+    // Firefox inlude entries in the HAR that are from the local cache
+    // and adds them with a time of 0.
+    // lets remove them.
+    remove(har.log.entries, function(entry) {
+      return entry.time === 0;
+    });
+    this.hars.push(har);
+    this.index++;
   }
 
-  onStopRun(result) {
+  async onStopRun(result) {
     if (!this.skipHar && this.hars.length > 0) {
       result.har = harBuilder.mergeHars(this.hars);
     }
-
-    return Promise.resolve();
   }
 }
 
