@@ -1,8 +1,8 @@
 'use strict';
 
-const ChromeDelegate = require('./chromeDelegate'),
-  FirefoxDelegate = require('./firefoxDelegate'),
-  nullDelegate = require('./nullDelegate');
+const ChromeDelegate = require('./chromeDelegate');
+const FirefoxDelegate = require('./firefoxDelegate');
+const nullDelegate = require('./nullDelegate');
 
 module.exports = {
   createDelegate(options) {
