@@ -1,9 +1,6 @@
 'use strict';
 
-const Promise = require('bluebird');
-
 const delegate = {};
-delegate.onStartRun = delegate.onStartIteration = delegate.onStopIteration = delegate.onStopRun =
-  Promise.resolve;
+delegate.onStartRun = delegate.onStartIteration = delegate.onStopIteration = delegate.onStopRun = async () => {};
 
 module.exports = delegate;
