@@ -12,7 +12,9 @@ Promise.promisifyAll(fs);
 const mkdirp = Promise.promisify(require('mkdirp'));
 
 const defaultDir = 'browsertime-results';
-let timestamp = moment().format().replace(/:/g, '');
+let timestamp = moment()
+  .format()
+  .replace(/:/g, '');
 
 function pathNameFromUrl(url) {
   const parsedUrl = urlParser.parse(url),
@@ -22,7 +24,10 @@ function pathNameFromUrl(url) {
 
   if (!isEmpty(parsedUrl.search)) {
     const md5 = crypto.createHash('md5'),
-      hash = md5.update(parsedUrl.search).digest('hex').substring(0, 8);
+      hash = md5
+        .update(parsedUrl.search)
+        .digest('hex')
+        .substring(0, 8);
     pathSegments.push('query-' + hash);
   }
 
