@@ -3,7 +3,9 @@
 const execa = require('execa'),
   path = require('path'),
   Promise = require('bluebird'),
-  fs = require('fs');
+  fs = require('fs'),
+  get = require('lodash.get'),
+  videoDefaults = require('../defaults');
 
 Promise.promisifyAll(fs);
 
@@ -15,12 +17,12 @@ module.exports = {
       '-i',
       videoFile,
       '-r',
-      context.options.videoParams.framerate,
+      get(context, 'options.videoParams.framerate', videoDefaults.framerate),
       tmpFile
     ];
     context.log.info(
       'Converting video to %s fps',
-      context.options.videoParams.framerate
+      get(context, 'options.videoParams.framerate', videoDefaults.framerate)
     );
     return execa('ffmpeg', scriptArgs).then(() => {
       return fs.renameAsync(tmpFile, videoFile);
