@@ -1,7 +1,9 @@
 'use strict';
 
 const execa = require('execa'),
-  log = require('intel');
+  log = require('intel'),
+  get = require('lodash.get'),
+  videoDefaults = require('../defaults');
 
 module.exports = {
   run(videoPath, newStart, outputFile, metrics, options) {
@@ -17,24 +19,34 @@ module.exports = {
       const speedIndex = `drawtext=${fontFile}enable='between(t,${Number(
         metrics.SpeedIndex
       ) /
-        1000},30)':x=(w-tw)/2: y=H-124:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='SpeedIndex ${metrics.SpeedIndex}'`;
+        1000},30)':x=(w-tw)/2: y=H-124:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='SpeedIndex ${
+        metrics.SpeedIndex
+      }'`;
 
       const firstVisualChange = `drawtext=${fontFile}enable='between(t,${Number(
         metrics.FirstVisualChange
       ) /
-        1000},30)':x=(w-tw)/2: y=H-180:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='FirstVisualChange ${metrics.FirstVisualChange}'`;
+        1000},30)':x=(w-tw)/2: y=H-180:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='FirstVisualChange ${
+        metrics.FirstVisualChange
+      }'`;
 
       const visualComplete85 = `drawtext=${fontFile}enable='between(t,${Number(
         metrics.VisualComplete85
       ) /
-        1000},30)':x=(w-tw)/2: y=H-152:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='VisualComplete85 ${metrics.VisualComplete85}'`;
+        1000},30)':x=(w-tw)/2: y=H-152:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='VisualComplete85 ${
+        metrics.VisualComplete85
+      }'`;
 
       const lastVisualChange = `drawtext=${fontFile}enable='between(t,${Number(
         metrics.LastVisualChange
       ) /
-        1000},30)':x=(w-tw)/2: y=H-96:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='LastVisualChange ${metrics.LastVisualChange}'`;
+        1000},30)':x=(w-tw)/2: y=H-96:fontcolor=white:fontsize=26:box=1:boxcolor=0x000000AA:text='LastVisualChange ${
+        metrics.LastVisualChange
+      }'`;
 
-      extras = `,${speedIndex}, ${lastVisualChange}, ${visualComplete85}, ${firstVisualChange}`;
+      extras = `,${speedIndex}, ${lastVisualChange}, ${visualComplete85}, ${
+        firstVisualChange
+      }`;
     }
 
     let scriptArgs = [
@@ -48,13 +60,22 @@ module.exports = {
     ];
 
     // keeping old legacy videoRaw
-    const showTimer = options.videoRaw ? false : options.videoParams.addTimer;
+    let showTimer = options.videoRaw ? false : videoDefaults.addTimer;
+    if (
+      !options.videoRaw &&
+      get(options, 'videoParams.addTimer', undefined) != undefined
+    ) {
+      showTimer = options.videoParams.addTimer;
+    }
+
     if (showTimer) {
       scriptArgs.push(
         '-vf',
-        `drawtext=${fontFile}timecode='00\\:00\\:00\\:00':rate=${options
-          .videoParams
-          .framerate}: x=(w-280)/2: y=H-60:fontcolor=white:fontsize=48:box=1:boxcolor=0x000000AA${extras}`
+        `drawtext=${fontFile}timecode='00\\:00\\:00\\:00':rate=${
+          options.videoParams.framerate
+        }: x=(w-280)/2: y=H-60:fontcolor=white:fontsize=48:box=1:boxcolor=0x000000AA${
+          extras
+        }`
       );
     }
 
