@@ -10,7 +10,7 @@ module.exports = {
       '-f',
       'lavfi',
       '-i',
-      "movie='" + videoPath + "',signalstats",
+      "movie='" + videoPath.replace('=', '\\=') + "',signalstats",
       '-show_entries',
       'frame_tags=lavfi.signalstats.SATAVG',
       '-v',
