@@ -50,10 +50,18 @@ module.exports = {
             if (clientId) {
               taskData.deviceid = clientId;
               context.log.info('Start screenrecord on device %s', clientId);
-              return client.shell(
-                clientId,
-                'screenrecord --bit-rate 8000000 /sdcard/browsertime.mp4'
-              );
+              return client
+                .shell(clientId, 'echo $EXTERNAL_STORAGE')
+                .then(adb.util.readAll)
+                .then(output => (taskData.sdcard = output.toString().trim()))
+                .then(() =>
+                  client.shell(
+                    clientId,
+                    'screenrecord --bit-rate 8000000 ' +
+                      taskData.sdcard +
+                      '/browsertime.mp4'
+                  )
+                );
             } else {
               return client.listDevices().then(devices => {
                 // just take the first phone online
@@ -62,10 +70,18 @@ module.exports = {
                   'Start screenrecord on device %s',
                   devices[0].id
                 );
-                return client.shell(
-                  devices[0].id,
-                  'screenrecord --bit-rate 8000000 /sdcard/browsertime.mp4'
-                );
+                return client
+                  .shell(devices[0].id, 'echo $EXTERNAL_STORAGE')
+                  .then(adb.util.readAll)
+                  .then(output => (taskData.sdcard = output.toString().trim()))
+                  .then(() =>
+                    client.shell(
+                      devices[0].id,
+                      'screenrecord --bit-rate 8000000 ' +
+                        taskData.sdcard +
+                        '/browsertime.mp4'
+                    )
+                  );
               });
             }
           })
